backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
lxml==5.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
